Update todos to use datatable, separate suggested todos from the rest.
